==> analysis/3_ClassificationModels.Rmd <==
---
title: "3_ClassificationModels"
author: "massaine"
date: "2022-11-30"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

### Accuracy of PT methods
```{r}
options(stringsAsFactors=FALSE)
library(prospectr)
library(caret)
library(mlbench)
library(pROC)
library(C50)
library(stringr)
library(dplyr)
library(gt)
library(ggpubr) 
```

```{r, eval=FALSE}
NIR <- "SCIO"
load(file=here::here("output","dadosSCIO_PT.RData"))

NIR <- "BUCHI"
load(file=here::here("output","dadosBUCHI_PT.RData"))
```

### test using PLS model
```{r, eval=F}
ctrl <- trainControl(method="repeatedcv",repeats = 5) 
```

### KNN model to test pre treatment methods
```{r, eval = F}
trait <- "Waxy"
modelFit2 <- list()
j=1
for (j in 1:9) {
  dados <- dados2[[j]]
  set.seed(2304)
  modelFit1 <- train(Waxy ~ ., data = dados, method = "knn",
                     trControl = ctrl,preProcess = c("center","scale"),
                     tuneLength = 1)
  modelFit2[[j]] <- modelFit1
  }

```

### results
```{r, eval = F}
res <- resamples(modelFit2)
res <- res$values
resCV_1 <- list()

for(i in 1:9){
  resCV_1[[i]]<-res[,c(paste(paste("Model",i,sep=""),"~Accuracy",sep=""),
                         paste(paste("Model",i,sep =""),"~Kappa",sep=""))]
  colnames(resCV_1[[i]]) <- c("Accuracy","Kappa")
  }

resCV_2 <- do.call(rbind,resCV_1)
resCV_2$PT <- rep(c(1:9),each=50)
resCV_2$PT <- as.factor(paste("pt",resCV_2$PT,sep = "_"))
levels(resCV_2$PT) <- c("raw data", "1st","DT","MSC","SNV","1st + DT",
                        "1st + MSC","DT + MSC", "1st + SG + DT")
```

### Save and load results 
```{r}
#PT_BUCHI <- resCV_2
#PT_SCIO <- resCV_2
#save(PT_SCIO,PT_BUCHI,file = here::here("output","Results_PreTreat.RData"))
load(file = here::here("output","Results_PreTreat.RData"))
```


### plot 
```{r}
p1 <- ggplot(PT_BUCHI, aes(x=PT, y=Accuracy, fill=PT)) + theme_bw() + 
  geom_boxplot(outlier.shape = NA) + 
  theme(legend.position = "none",axis.title.x=element_blank())+
  ylim(0.8, 1) + ggtitle("A) NIR BUCHI")
 
p2<- ggplot(PT_BUCHI, aes(x=PT, y=Kappa, fill=PT)) +theme_bw() + 
  geom_boxplot(outlier.shape = NA) + 
  theme(legend.position = "none",axis.title.x=element_blank())+
  ylim(0, 1) + ggtitle("A) NIR BUCHI")

p3<- ggplot(PT_SCIO, aes(x=PT, y=Accuracy, fill=PT)) + theme_bw() + 
  geom_boxplot(outlier.shape = NA) + 
  theme(legend.position = "none",axis.title.x=element_blank())+
  ylim(.8, 1) + ggtitle("B) NIR SCIO")
 
p4<- ggplot(PT_SCIO, aes(x=PT, y=Kappa, fill=PT)) + theme_bw() + 
  geom_boxplot(outlier.shape = NA) + 
  theme(legend.position = "none",axis.title.x=element_blank())+
  ylim(0, 1) + ggtitle("B) NIR SCIO")


tiff(file = "Figure1_PT_methods_Acc.tiff", 
               width = 19, height=12, res = 700, 
               compression = "lzw", units = "cm")
ggarrange(p1,p3,ncol=1,nrow=2)
dev.off()

tiff(file = "Figure1_PT_methods_Kap.tiff", 
               width = 19, height=12, res = 700, 
               compression = "lzw", units = "cm")
ggarrange(p2,p4,ncol=1,nrow=4)
dev.off()
```

### table BUCHI
```{r}
 PT_BUCHI%>%group_by(PT)%>%
   summarise(Accm = round(mean(Accuracy, na.rm = T),3),
            Acc.sd = round(sd(Accuracy, na.rm = T),3),
            Kap.m = round(mean(Kappa, na.rm = T),3),
            Kap.sd = round(sd(Kappa, na.rm = T),3)) %>%
       gt()  %>%     tab_options(table.font.size = pct(75))
```

### table SCIO
```{r}
 PT_SCIO%>%group_by(PT)%>%
   summarise(Accm = round(mean(Accuracy, na.rm = T),3),
            Acc.sd = round(sd(Accuracy, na.rm = T),3),
            Kap.m = round(mean(Kappa, na.rm = T),3),
            Kap.sd = round(sd(Kappa, na.rm = T),3)) %>%
       gt()  %>%     tab_options(table.font.size = pct(75))

```
