==> analysis/4.4_RF_SCIO_ClassificationModels.Rmd <==
---
title: "3_ClassificationModels"
author: "massaine"
date: "2022-11-30"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

## Introduction

### Accuracy of PT methods
```{r}
options(stringsAsFactors=FALSE)
library(prospectr)
library(caret)
library(mlbench)
library(pROC)
library(C50)
library(stringr)
library(dplyr)
library(gt)
```

```{r, eval=FALSE}
NIR <- "SCIO"
load(file=here::here("output","dadosSCIO_PT.RData"))

#NIR <- "BUCHI"
#load(file=here::here("output","dadosBUCHI_PT.RData"))
```

### test using PLS model
```{r}
ctrl <- trainControl(method="repeatedcv",repeats = 5) 
```


### KNN model to test pre treatment methods
#### Select the best PT method (1st + MSC)
```{r,eval=F}
dados1 <- dados2[[7]]
dados1$Waxy <- as.factor(dados1$Waxy)
#modelsnames <- c("knn", "C5.0", "xgbLinear","parRF")
modelsnames <- c("parRF")
modeln <- "parRF"
model <- list()
#n <- length(modelsnames)
n=1
modelFit2 <- list()
pCM <- list()

#for (j in 1:n) {
  modelFit <- list()
  index <- list()
  training <- list()
  testing <- list()
  CM <- list()

  i=1
for (i in 1:5) {
   set.seed(i)
    index[[i]] <- createDataPartition(dados1[,"Waxy"], p = .8, list = FALSE)
    training[[i]] <- dados1[index[[i]], ]
    testing[[i]] <- dados1[-index[[i]], ]
    
    set.seed(i)
    modelFit[[i]] <- train(Waxy ~ ., data = training[[i]], 
                           method = modelsnames[1], 
                           trControl = ctrl,
                           preProcess = c("center","scale")) 
    
    modelPredict <- predict(modelFit[[i]],newdata = testing[[i]])
    CM[[i]] <- confusionMatrix(modelPredict, testing[[i]]$Waxy)
}
 # modelFit2[[j]] <- modelFit
 #  pCM[[j]] <- CM 
  #}

if(NIR=="SCIO")
{save(CM, modelFit,file =here::here("output",
                   paste(modeln,"SCIO_Model.RData",sep ="_")))}
if(NIR=="SCIO")
{save(training,testing,file =here::here("output","SCIO_testing_set.RData"))}
  
if(NIR=="BUCHI")
{save(CM, modelFit,file =here::here("output",
                   paste(modeln,"BUCHI_Model.RData",sep ="_")))}

```
