==> analysis/6_Selecao_Variaveis.Rmd <==
---
title: "6_Selecao de Variaveis"
author: "massaine"
date: "2022-11-30"
output: workflowr::wflow_html
editor_options:
  chunk_output_type: console
---

### Function Confusion matrix
```{r}
prettyConfused<-function(Actual,Predict,colors=c("white","red4",
                                     "dodgerblue3"),text.scl=5){
  actual = as.data.frame(table(Actual))
  names(actual) = c("Actual","ActualFreq")
  #build confusion matrix
  confusion = as.data.frame(table(Actual, Predict))
  names(confusion) = c("Actual","Predicted","Freq")
  #calculate percentage of test cases based on actual frequency
  confusion = merge(confusion, actual, by=c('Actual','Actual'))
  confusion$Percent = confusion$Freq/confusion$ActualFreq*100
  confusion$ColorScale<-confusion$Percent*-1
  confusion[which(confusion$Actual==confusion$Predicted),]$ColorScale<-confusion[which(confusion$Actual==confusion$Predicted),]$ColorScale*-1
  confusion$Label<-paste(round(confusion$Percent,0),"%, n=",
                         confusion$Freq,sep="")
  tile <- ggplot() + theme(panel.grid.major = element_blank(),
                           panel.grid.minor = element_blank(),
                          panel.background = element_blank(), 
                          axis.line = element_blank(),
                          axis.title.x = element_text(size = rel(1)),
                          axis.title.y = element_text(size = rel(1)),
                          axis.text.x = element_text(size=13),
                          axis.text.y = element_text(size=13)) +
              geom_tile(aes(x=Actual, y=Predicted,fill=ColorScale),
              data=confusion, color="black",size=0.1) +
              labs(x="Reference",y="Predicted")
  
  tile = tile + geom_text(aes(x=Actual,y=Predicted, label=Label),
              data=confusion, size=text.scl, colour="black") +
               scale_fill_gradient2(low=colors[2],
               high=colors[3],
               mid=colors[1],midpoint = 0,guide='none')
}
library(caret)
```

### Load results from classification models (SCIO)
```{r,eval=F}
load(file=here::here("output","xgbLinear_SCIO_Model.RData"))
load(file=here::here("output","SCIO_testing_set.RData"))

modelFit_scio <- modelFit[[1]]
Imp <- varImp(modelFit_scio, scale = FALSE)
plot(Imp, top = 30)
impMat <- as.data.frame(Imp$importance)
names <- str_split_fixed(rownames(impMat), "`", 3)
impMat$Waves <- names[,2]
impMat$Overall
rownames(impMat)<-seq(1:nrow(impMat))
head(impMat)
impMat$perc[1] <- 100
for (i in 2:nrow(impMat)) {impMat$perc[i] <- (impMat[i,1]*100)/impMat[1,1] }

plotimpS <- ggplot(data= impMat, aes(x=as.numeric(Waves), y=perc)) +
          geom_point() + theme_bw() + theme(legend.title = element_blank(),
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(),
          panel.background = element_blank(), axis.line = element_blank()) +
          geom_hline(yintercept = 30,colour ="red") + ggtitle("SCiO") +
          scale_y_continuous("Variable importance (%)",
          breaks = seq(0,100,20)) + scale_x_continuous("Wavelengths")

plotimpS
```

### Classification models after Var. Selection (100 waveslength)
```{r, eval=FALSE}
ctrl <- trainControl(method="repeatedcv",repeats = 5) 
sel <- as.numeric(impMat$Waves[1:34])                 # 30%
sel1 <- as.character(sel[order(sel,decreasing = F)])
trainingSCIO <- training[[1]][,c("Waxy",sel1)]
testingSCIO <- testing[[1]]
testingSCIO <- testingSCIO[,c("Waxy",sel1)]
levels(testingSCIO$Waxy) <- list(Nonwaxy = "No", Waxy = "Waxy")

set.seed(1)
modelFitSEL_S1 <- train(Waxy ~ ., data = trainingSCIO, 
                        method = "knn", trControl = ctrl,           
                      preProcess = c("center","scale"))
set.seed(1)
modelFitSEL_S2 <- train(Waxy ~ ., data = trainingSCIO, 
                        method = "C5.0", trControl = ctrl,           
                      preProcess = c("center","scale"))
set.seed(1)
modelFitSEL_S3 <- train(Waxy ~ ., data = trainingSCIO, 
                        method = "xgbLinear", trControl = ctrl,           
                      preProcess = c("center","scale"))
set.seed(1)
modelFitSELS_S4 <- train(Waxy ~ ., data = trainingSCIO, 
                         method = "parRF", trControl = ctrl,           
                      preProcess = c("center","scale"))

#save(modelFitSEL_S1,modelFitSEL_S2,modelFitSEL_S3,modelFitSEL_S4,
#     file = here::here("output","SCIO_Model_Sel_Var.RData"))
```


### Load results from classification models (BUCHI)
```{r,eval=F}
load(file=here::here("output","xgbLinear_BUCHI_Model.RData"))
load(file=here::here("output","BUCHI_testing_set.RData"))

modelFit_BUCHI <- modelFit[[1]]
Imp <- varImp(modelFit_BUCHI, scale = FALSE)
plot(Imp, top = 30)
impMat <- as.data.frame(Imp$importance)
names <- str_split_fixed(rownames(impMat), "`", 3)
impMat$Waves <- names[,2]
impMat$Overall
rownames(impMat)<-seq(1:nrow(impMat))
head(impMat)
impMat$perc[1] <- 100
for (i in 2:nrow(impMat)) {impMat$perc[i] <- (impMat[i,1]*100)/impMat[1,1] }

plotimpB <- ggplot(data= impMat, aes(x=as.numeric(Waves), y=perc)) +
          geom_point() + theme_bw() + theme(legend.title = element_blank(),
          panel.grid.major = element_blank(), 
          panel.grid.minor = element_blank(),
          panel.background = element_blank(), axis.line = element_blank()) +
          geom_hline(yintercept = 30,colour ="red") + 
          ggtitle("NIRFlex N-500") +
          scale_y_continuous("Variable importance (%)",
          breaks = seq(0,100,20)) + scale_x_continuous("Wavelengths")

plotimpB
```

### Classification models after Var. Selection (100 waveslength)
```{r,eval=FALSE}
ctrl <- trainControl(method="repeatedcv",repeats = 5) 
sel <- as.numeric(impMat$Waves[1:37])                 # 30%
sel1 <- as.character(sel[order(sel,decreasing = F)])
trainingBUCHI <- training[[1]][,c("Waxy",sel1)]
testingBUCHI <- testing[[1]]
testingBUCHI <- testingBUCHI[,c("Waxy",sel1)]
levels(testingBUCHI$Waxy) <- list(Nonwaxy = "No", Waxy = "Waxy")

set.seed(1)
modelFitSEL_S1 <- train(Waxy ~ ., data = trainingBUCHI, 
                        method = "knn", trControl = ctrl,           
                      preProcess = c("center","scale"))
set.seed(1)
modelFitSEL_S2 <- train(Waxy ~ ., data = trainingBUCHI, 
                        method = "C5.0", trControl = ctrl,           
                      preProcess = c("center","scale"))
set.seed(1)
modelFitSEL_S3 <- train(Waxy ~ ., data = trainingBUCHI, 
                        method = "xgbLinear", trControl = ctrl,           
                      preProcess = c("center","scale"))
set.seed(1)
modelFitSEL_S4 <- train(Waxy ~ ., data = trainingBUCHI, 
                         method = "parRF", trControl = ctrl,           
                      preProcess = c("center","scale"))

save(modelFitSEL_S1,modelFitSEL_S2,modelFitSEL_S3,modelFitSEL_S4,
     file = here::here("output","BUCHI_Model_Sel_Var.RData"))

```

### Plot
```{r}
tiff(file = "Figure3_SelVariables.tiff", 
               width = 14, height=8, res = 700, 
               compression = "lzw", units = "cm")
ggarrange(plotimpS,plotimpB,ncol=2,nrow=1)
dev.off()
```
